typecast to double to fix build error

+++ src/esp/gfx/ssao/nvpro_core/nvh/cameramanipulator.cpp
@@ -530,10 +530,10 @@ void CameraManipulator::fit(const nvmath::vec3f& boxMin,
       {
         // Keep the largest offset to see that vertex
         offset =
-            std::max(fabs(vct.y) / tan(nv_to_rad * yfov * 0.5f) + fabs(vct.z),
+            std::max(double(fabs(vct.y)) / tan(nv_to_rad * yfov * 0.5f) + fabs(vct.z),
                      double(offset));
         offset =
-            std::max(fabs(vct.x) / tan(nv_to_rad * xfov * 0.5f) + fabs(vct.z),
+            std::max(double(fabs(vct.x)) / tan(nv_to_rad * xfov * 0.5f) + fabs(vct.z),
                      double(offset));
       }
     }
